fix(fixtures): Move python files to /fixtures/src/
Move python source files to /fixtures/src and fix imports
