+++ fixtures/src/queries/subjects.py
